veritas/requirements.txt
# Django core
Django>=5.1,<6.0
psycopg[binary]>=3.1

# REST API
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

# Environment / config
django-environ>=0.11

# Structured logging
django-structlog>=8.0
structlog>=24.0

# Celery + Redis
celery[redis]>=5.4
redis>=5.0

# pgvector (Phase 2+ embeddings, but enabling extension early)
pgvector>=0.3

# CORS (for frontend)
django-cors-headers>=4.3

# Gunicorn for production
gunicorn>=22.0

# Code quality
ruff>=0.4
